Refactor: reorganize projects
