+++ src/webapiProject/DataAccess/Interfaces/IExampleRepository.cs
@@ -1,0 +1,5 @@
+﻿namespace webapiProject.DataAccess.Interfaces;
+
+public interface IExampleRepository
+{
+}
