+++ src/webapiProject/DataAccess/Repositories/ExampleRepository.cs
@@ -1,0 +1,7 @@
+﻿using webapiProject.DataAccess.Interfaces;
+
+namespace webapiProject.DataAccess.Repositories;
+
+public class ExampleRepository : IExampleRepository
+{
+}
