+++ src/webapiProject/Program.cs
@@ -1,0 +1,61 @@
+using Microsoft.OpenApi.Models;
+using System.Reflection;
+using webapiProject.Core.Interfaces;
+using webapiProject.Core.Services;
+
+var builder = WebApplication.CreateBuilder(args);
+
+
+
+builder.Services.AddControllers();
+// Learn more about configuring Swagger/OpenAPI at https://aka.ms/aspnetcore/swashbuckle
+builder.Services.AddEndpointsApiExplorer();
+builder.Services.AddSwaggerGen();
+builder.Services.AddSwaggerGen(c =>
+{
+    //The generated Swagger JSON file will have these properties.
+    //Update Api info:
+    c.SwaggerDoc("v1", new OpenApiInfo
+    {
+        Title = "Your Api Name",
+        Version = "v1",
+        Description = "Your Api Description.",
+        TermsOfService = new Uri("http://www.yourUrl.com/TermsOfService"),
+        Contact = new OpenApiContact
+        { Name = "Contact Title", Email = "contactemailaddress@domain.com", Url = new Uri("http://www.yourUrl.com") }
+    });
+
+
+    //Locate the XML file being generated by ASP.NET...
+    var xmlFile = $"{Assembly.GetExecutingAssembly().GetName().Name}.XML";
+    var xmlPath = Path.Combine(AppContext.BaseDirectory, xmlFile);
+    //... and tell Swagger to use those XML comments.
+    c.IncludeXmlComments(xmlPath);
+});
+
+// Add services to the container.
+builder.Services.AddScoped<IExampleService, ExampleService>();
+
+// Add repositories to services
+//builder.Services.AddScoped<IExampleRepository, Examplerepository>();
+//builder.Services.AddScoped<IUnitOfWork, UnitOfWork>();
+
+//Configure automapper dependency injection
+builder.Services.AddAutoMapper(AppDomain.CurrentDomain.GetAssemblies());
+
+var app = builder.Build();
+
+// Configure the HTTP request pipeline.
+if (app.Environment.IsDevelopment())
+{
+    app.UseSwagger();
+    app.UseSwaggerUI();
+}
+
+app.UseHttpsRedirection();
+
+app.UseAuthorization();
+
+app.MapControllers();
+
+app.Run();
