--- webapiProject/DataAccess/Interfaces/IExampleRepository.cs
@@ -1,5 +1,0 @@
-﻿namespace webapiProject.DataAccess.Interfaces;
-
-public interface IExampleRepository
-{
-}
