--- webapiProject/DataAccess/Repositories/ExampleRepository.cs
@@ -1,7 +1,0 @@
-﻿using webapiProject.DataAccess.Interfaces;
-
-namespace webapiProject.DataAccess.Repositories;
-
-public class ExampleRepository : IExampleRepository
-{
-}
